Refactor: fix packages and imports to new design pattern
