+++ grails-app/services/com/mini/asaas/AddressService.groovy
@@ -1,8 +1,9 @@
 package com.mini.asaas
 
 import com.mini.asaas.Address
-import grails.gorm.transactions.Transactional
 import com.mini.asaas.dto.AddressDTO
+
+import grails.gorm.transactions.Transactional
 
 @Transactional
 class AddressService {
