+++ grails-app/services/com/mini/asaas/CustomerService.groovy
@@ -3,8 +3,10 @@ package com.mini.asaas
 import com.mini.asaas.Customer
 import com.mini.asaas.Address
 import com.mini.asaas.AddressService
-import grails.gorm.transactions.Transactional
 import com.mini.asaas.dto.CustomerDTO
+import com.mini.asaas.enums.PersonType
+
+import grails.gorm.transactions.Transactional
 
 @Transactional
 class CustomerService {
