+++ grails-app/services/com/mini/asaas/PayerService.groovy
@@ -1,14 +1,17 @@
 package com.mini.asaas
 
-import javax.transaction.Transactional
 import com.mini.asaas.Payer
 import com.mini.asaas.Address
-import com.mini.asaas.AddressService
 import com.mini.asaas.Customer
+import com.mini.asaas.AddressService
 import com.mini.asaas.dto.PayerDTO
+import com.mini.asaas.enums.PersonType
+
+import javax.transaction.Transactional
 
 @Transactional
 class PayerService {
+    
     AddressService addressService
 
     public Payer save(PayerDTO payerDTO, Long customerId) {
