+++ src/main/groovy/com/mini/asaas/utils/BaseEntity.groovy
